stripe checkout

+++ db/migrate/20210605201458_create_products.rb
@@ -4,7 +4,7 @@ class CreateProducts < ActiveRecord::Migration[6.1]
       t.string :name
       t.string :desc
       t.integer :SKU
-      t.decimal :price_cents
+      t.integer :price_cents
       t.references :store
 
       t.timestamps
